Rework way we add storage-systems for monitoring
The prior way that we were allowing storage-systems to be added for
management wasn't very easy to utilize. It was difficult to understand and
didn't scale very well. Rather than utilizing Ansible, this new method
integrates with the collector script and allows a single JSON
configuration file to be provided with storage-systems.

This patch adds functionality to the collector.py script to utilize
config.json for the systems to monitor. The collector script will add the
systems automatically. The documentation has been updated to reflect this
and a sample file has been provided.

Additionally, a small addition has been made to the Overview dashboard to
make it a bit more clear on what dashboards we provide and indicate why
there may be no storage-systems shown.